requirements.txt
scrapy==2.8.0
beautifulsoup4==4.11.2
PyPDF2==3.0.1
openai==0.27.8
schedule==1.2.0
requests==2.31.0
python-dotenv==1.0.0
anthropic~=0.30.1